Split DesignerView into VariableView.
